Refactor initiation protocols into packet handlers
