+++ src/game_state.rs
@@ -2,9 +2,9 @@ pub mod block;
 pub mod patchwork;
 pub mod player;
 
-use super::gameplay_router;
 use super::map;
 use super::messenger;
 use super::packet;
+use super::packet_handlers::gameplay_router;
 use super::packet_processor;
 use super::server;
