--- src/initiation_protocols.rs
@@ -1,9 +1,0 @@
-pub mod border_cross_login;
-pub mod client_ping;
-pub mod handshake;
-pub mod login;
-pub mod peer_subscription;
-
-use super::game_state;
-use super::messenger;
-use super::packet;
